backgrounds are now properly backed up, so their transitions work again.
slide transitions are still broken because the layout() code collides with the CSS styles: those are overruled by the layout() positioning code.

proper slide dimensioning for layout, including config option where all slides will have the same scale for a unified look (but not in the overview mode as there you want each text to be readable or at least attempt to, so there it's every slide for himself).